Refactor: Update imports and cleanup files after architecture changes
This commit addresses the changes required after a project refactoring to a clean architecture pattern.

Key changes include:
- Updated all Python import statements throughout `backend/src/app` and `backend/tests` to reflect new module paths:
    - `app.core.config` → `app.config`
    - `app.core.security` → `app.security` (restored missing file)
    - `app.core.db` → `app.adapters.orm`
    - `app.models` → `app.domain.user`, `app.domain.item`, `app.adapters.orm`, or `app.entrypoints.schemas`
    - `app.main` → `app.entrypoints.fastapi_app`
    - `app.api.deps` → `app.entrypoints.api.deps`
    - `app.api.routes.*` → `app.entrypoints.api.routes.*`
- Updated configuration files (`backend/.pyre_configuration`, `backend/Dockerfile`, `scripts/generate-client.sh`, `backend/README.md`) to reflect new paths and module locations.
- Removed obsolete files and directories:
    - `backend/src/app/core/`
    - `backend/src/app/api/`
    - `backend/src/app/main.py`
    - `backend/src/app/models/` was confirmed to be empty/non-existent.

Verification:
- Python imports are now resolving correctly.
- A separate issue with database connectivity during tests (`sqlalchemy.exc.OperationalError: (psycopg.OperationalError) [Errno -2] Name or service not known`) was identified but is considered outside the scope of this import/file structure refactoring.
