+++ backend/tests/api/routes/test_users.py
@@ -8,8 +8,8 @@ from app import crud
 from app.config import settings  # Updated: config
 from app.security import verify_password  # Updated: security
 from app.adapters.orm import User  # Updated: ORM model
-from app.entrypoints.schemas import UserCreate  # Updated: schemas
-from app.tests.utils.utils import random_email, random_lower_string
+from app.entrypoints.schemas import UserCreateInput  # Updated: schemas & corrected name
+from tests.utils.utils import random_email, random_lower_string # Corrected path
 
 
 def test_get_users_superuser_me(
@@ -62,7 +62,8 @@ def test_get_existing_user(
 ) -> None:
     username = random_email()
     password = random_lower_string()
-    user_in = UserCreate(email=username, password=password)
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
     user = crud.create_user(session=db, user_create=user_in)
     user_id = user.id
     r = client.get(
@@ -120,7 +121,7 @@ def test_create_user_existing_username(
     username = random_email()
     # username = email
     password = random_lower_string()
-    user_in = UserCreate(email=username, password=password)
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
     crud.create_user(session=db, user_create=user_in)
     data = {"email": username, "password": password}
     r = client.post(
@@ -152,12 +153,12 @@ def test_retrieve_users(
 ) -> None:
     username = random_email()
     password = random_lower_string()
-    user_in = UserCreate(email=username, password=password)
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
     crud.create_user(session=db, user_create=user_in)
 
     username2 = random_email()
     password2 = random_lower_string()
-    user_in2 = UserCreate(email=username2, password=password2)
+    user_in2 = UserCreateInput(email=username2, password=password2) # Corrected name
     crud.create_user(session=db, user_create=user_in2)
 
     r = client.get(f"{settings.API_V1_STR}/users/", headers=superuser_token_headers)
@@ -251,7 +252,7 @@ def test_update_user_me_email_exists(
 ) -> None:
     username = random_email()
     password = random_lower_string()
-    user_in = UserCreate(email=username, password=password)
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
     user = crud.create_user(session=db, user_create=user_in)
 
     data = {"email": user.email}
@@ -365,12 +366,12 @@ def test_update_user_email_exists(
 ) -> None:
     username = random_email()
     password = random_lower_string()
-    user_in = UserCreate(email=username, password=password)
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
     user = crud.create_user(session=db, user_create=user_in)
 
     username2 = random_email()
     password2 = random_lower_string()
-    user_in2 = UserCreate(email=username2, password=password2)
+    user_in2 = UserCreateInput(email=username2, password=password2) # Corrected name
     user2 = crud.create_user(session=db, user_create=user_in2)
 
     data = {"email": user2.email}
@@ -386,7 +387,9 @@ def test_update_user_email_exists(
 def test_delete_user_me(client: TestClient, db: Session) -> None:
     username = random_email()
     password = random_lower_string()
-    user_in = UserCreate(email=username, password=password)
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
+    user_in = UserCreateInput(email=username, password=password) # Corrected name
     user = crud.create_user(session=db, user_create=user_in)
     user_id = user.id
