+++ backend/tests/utils/item.py
@@ -1,9 +1,10 @@
 from sqlmodel import Session
 
 from app import crud
-from app.models import Item, ItemCreate
-from app.tests.utils.user import create_random_user
-from app.tests.utils.utils import random_lower_string
+from app.adapters.orm import Item
+from app.entrypoints.schemas import ItemCreateInput # Corrected name
+from tests.utils.user import create_random_user # Corrected path
+from tests.utils.utils import random_lower_string # Corrected path
 
 
 def create_random_item(db: Session) -> Item:
@@ -12,5 +13,5 @@ def create_random_item(db: Session) -> Item:
     assert owner_id is not None
     title = random_lower_string()
     description = random_lower_string()
-    item_in = ItemCreate(title=title, description=description)
+    item_in = ItemCreateInput(title=title, description=description) # Corrected name
     return crud.create_item(session=db, item_in=item_in, owner_id=owner_id)
